gcc warn useless conversion but msvc gives an error

+++ include/nlohmann/detail/input/binary_reader.hpp
@@ -2072,7 +2072,7 @@ class binary_reader
                             {
                                 return false;
                             }
-                            result = number;
+                            result = static_cast<std::size_t>(number);
                             return true;
                         }
                         case '[':
